requirements.txt
streamlit
plotly
pandas
nltk
pyttsx3
wikipedia-api
SpeechRecognition
pyaudio
